Added method to create board and rooms
the board object gets created. the xml parser reads/collects/creates each room. all the rooms are stored in an array list, which is stored in the board object. Each room object has an arraylist of strings with the name of neighbors. room object has an arraylist of roles with the role data for that room.